Change solar energy config from long to int

+++ src/main/scala/resonantinduction/core/Settings.java
@@ -71,7 +71,7 @@ public class Settings
     @Config(category = "Power", key = "Water_tubine_Ratio")
     public static int WATER_POWER_RATIO = 1;
     @Config(category = "Power", key = "Solor_Panel")
-    public static long SOLAR_ENERGY = 50;
+    public static int SOLAR_ENERGY = 50;
 
     public static void setModMetadata(ModMetadata metadata, String id, String name)
     {
